+++ packages/implementations/react/src/atomic/Divider/src/divider.types.ts
@@ -1,31 +1,12 @@
-import { HTMLAttributes, ReactNode } from "react";
-// import { DividerSprinkle } from "@rarui/styles";
+import { HTMLAttributes } from "react";
+import { DividerSprinkle } from "@rarui/styles";
 
-export interface DividerProperties {
+export interface DividerProperties extends DividerSprinkle {
   /**
-   * The content of the button.
-   * @TJS-type React.ReactNode
+   * The direction of the Divider.
    */
-  children?: ReactNode;
-  /**
-   * Change the direction of the divider.
-   * @default large
-   */
-  direction?: "horizontal" | "vertical";
-  /**
-   * Change the width of the divider.
-   * @default medium
-   */
-  width?: "medium" | "small";
-  /**
-   * Change the size of the divider.
-   * @default md
-   */
-  size?: "xs" | "s" | "md" | "lg" | "xl" | "xxl" | "xlg" | "xxlg";
-  /**
-   * Change the style of the divider.
-   */
-  style?: "solid" | "dashed";
+  direction?: "vertical" | "horizontal";
 }
 
-export type DividerProps = DividerProperties & HTMLAttributes<HTMLDivElement>;
+export type DividerProps = DividerProperties &
+  Omit<HTMLAttributes<HTMLDivElement>, "type">;