+++ packages/implementations/react/src/atomic/Divider/CHANGELOG.md
@@ -6,7 +6,7 @@ A Divider is a thin line used to separate or group content into lists and layout
 
 #### 🎉 New features
 
-- Added `children`, `direction`, `width`, `size` and `style` properties to the component.([#28](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/28) by [@luiza.rodrigues](https://git.rarolabs.com.br/luiza.rodrigues))
+- Added `size`, `type`, `thickness` and `direction` properties to the component.([#28](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/28) by [@luiza.rodrigues](https://git.rarolabs.com.br/luiza.rodrigues))
 - Added stories on Component. ([#28](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/28) by [@luiza.rodrigues](https://git.rarolabs.com.br/luiza.rodrigues))
 
 <!-- #### 🛠 Breaking changes -->